Replaced "echo ... && exit" with die().

+++ usr/lib/outlogger.sh
@@ -7,32 +7,40 @@ __logpipe=""
 __logsredirected=false
 __logredirfile=""
 
+function die()
+{
+    ${__logsredirected} && end_log_redirect
+    echo "ERROR! ${1}" >&2
+    exit 1
+}
+
 # Initialise logging by creating an empty folder
 # $1 should contain the base path for log files
 function init_logging()
 {
-	[[ -n "${__logpath}" ]] && echo "ERROR! Logging already initialised." >&2 && exit 127
-	[[ -z "${1}" ]] && echo "ERROR! A valid path must be supplied to initialise logging." >&2 && exit 127
+	[[ -n "${__logpath}" ]] && die "Logging already initialised."
+	[[ -z "${1}" ]] && die "A valid path must be supplied to init_logging()"
 
 	__logpath="${1}"
-	mkdir -p "${__logpath}"
-	[[ ! -w ${__logpath} ]] && echo "ERROR! Unable to write to ${__logpath} in init_logging()" >&2 && exit 127
+	mkdir -p "${__logpath}" 2>/dev/null || die "Unable to create ${__logpath} in init_logging()" 
+	[[ ! -w ${__logpath} ]] && die "Unable to write to ${__logpath} in init_logging()"
 
-	rm "${__logpath}"/* -f
+	rm "${__logpath}"/* -f || die "Unable to remove all files in ${__logpath} in init_logging()"
 }
 
 # Redirect stdout and stderr to a file.
 # $1 should contain the name of the file to log to 
 function redirect_output_to_file()
 {
-	[[ -z "${1}" ]] && echo "ERROR! log file name required when calling redirect_output_to_file()" >&2 && exit 127
-	${__logsredirected} && echo "ERROR! logs already redirected in redirect_output_to_file()" >&2 && exit 127
+	[[ -z "${1}" ]] && die "log file name required when calling redirect_output_to_file()"
+	${__logsredirected} && die "logs already redirected in redirect_output_to_file()"
+	[[ -z "${__logpath}" ]] && die "Logging not initialised in redirect_output_to_file() - you must call init_logging() first"
 	
 	# Store the current stdout and stderr file descriptors and redirect both to the log.
 	__logredirfile="${1}"
 	exec 5<&1 6<&2
-	exec 1> "${__logpath}/${__logredirfile}"
-	exec 2> "${__logpath}/${__logredirfile}"
+	exec 1>> "${__logpath}/${__logredirfile}"
+	exec 2>> "${__logpath}/${__logredirfile}"
 	
 	__logsredirected=true
 }
@@ -41,8 +49,9 @@ function redirect_output_to_file()
 # $1 should contain the name of the file to log to 
 function tee_output_to_file_stdout()
 {
-	[[ -z "${1}" ]] && echo "ERROR! log file name required when calling tee_output_to_file_stdout()" >&2 && exit 127
-	${__logsredirected} && echo "ERROR! logs already redirected in tee_output_to_file_stdout()" >&2 && exit 127
+	[[ -z "${1}" ]] && die "log file name required when calling tee_output_to_file_stdout()"
+	${__logsredirected} && die "logs already redirected in tee_output_to_file_stdout()"
+    [[ -z "${__logpath}" ]] && die "Logging not initialised in tee_output_to_file_stdout() - you must call init_logging() first"
 
 	# Store the current stdout file descriptor.
 	exec 5<&1 6<&2
@@ -55,27 +64,28 @@ function tee_output_to_file_stdout()
 	tee < "${__logpipe}" "${__logpath}/${__logredirfile}" &
 	
 	# Redirect stdout and stderr to the pipe
-	exec 1> "${__logpipe}"
-	exec 2> "${__logpipe}"
+	exec 1>> "${__logpipe}"
+	exec 2>> "${__logpipe}"
 	
 	__logsredirected=true
 }
 
 # Execute a command and log stdout and stderr to files
 # $1 should contain the base file name
-# $2 should contain the command to execute
+# $2+ should contain the command to execute and any parameters
 # If the log files are zero size they are deleted, any remaining log files are
 # added to __logfiles list.
 function exec_and_log()
 {
-	[[ -z "${__logpath}" ]] && echo "ERROR! Logging not initialised when calling exec_and_log()." >&2 && exit 127
-	[[ -z "${1}" ]] && echo "ERROR! A file name is expected when calling exec_and_log()." >&2 && exit 127
-	[[ -z "${2}" ]] && echo "ERROR! A command is expected when calling exec_and_log()." >&2 && exit 127
+	[[ -z "${1}" ]] && die "A file name is expected when calling exec_and_log()"
+	[[ -z "${2}" ]] && die "A command is expected when calling exec_and_log()"
+    [[ -z "${__logpath}" ]] && die "Logging not initialised when calling exec_and_log() - you must call init_logging() first"
 
 	l1="${__logpath}/${1}.out.log"
 	l2="${__logpath}/${1}.err.log"
-
-	eval "${2}" 1>> "$l1" 2>> "$l2"
+	
+	shift
+	eval "${@}" 1>> "$l1" 2>> "$l2"
 	rs=$?
 
 	[[ ! -s ${l1} ]] && rm "${l1}" -f
@@ -91,7 +101,8 @@ function exec_and_log()
 # $1 should contain the base file name
 function add_aux_logs()
 {
-	[[ -z "${__logpath}" ]] && echo "ERROR! Logging not initialised when calling add_aux_logs()." >&2 && exit 127
+    [[ -z "${1}" ]] && die "A file name is expected when calling add_aux_logs()"
+	[[ -z "${__logpath}" ]] && die "Logging not initialised when calling add_aux_logs() - you must call init_logging() first"
 	
 	logs=( "${__logpath}"/"${1}".*.aux.*.log )
 	for l in "${logs[@]}"
@@ -104,7 +115,7 @@ function add_aux_logs()
 # End the redirection of logging configured using redirect_output_to_file() or tee_output_to_file_stdout()
 function end_log_redirect()
 { 
-	if ! ${__logsredirected} ; then echo "ERROR! logs not redirected when calling end_log_redirect()." >&2 && exit 127 ; fi
+	${__logsredirected} || die "logs not redirected when calling end_log_redirect()"
 
 	# End logging redirect by restoring stdout and stderr and killing the pipe if there is one
 	[[ -n "${__logpipe}" ]] && rm -f "${__logpipe}"
@@ -118,8 +129,9 @@ function end_log_redirect()
 # $1 should contain the minimum size of the log before bzip will be used
 function bzip_large_logs()
 {
-	[[ -z "${__logpath}" ]] && echo "ERROR! Logging not initialised when calling bzip_large_logs()." >&2 && exit 127
-	if ${__logszipped} ; then echo "ERROR! Logs already bzipped when calling bzip_large_logs()." >&2 && exit 127 ; fi
+    (( "${1}" == 0 )) && die "A minimum size (in bytes) is required when calling bzip_large_logs()"
+    [[ -z "${__logpath}" ]] && die "Logging not initialised when calling bzip_large_logs()"
+	${__logszipped} && die "Logs already bzipped when calling bzip_large_logs()"
 
 	nlf=""
 	for lf in ${__logfiles}
@@ -135,7 +147,7 @@ function bzip_large_logs()
 # Returns a list of log files in $1
 function get_log_files()
 {
-	[[ -z "${__logpath}" ]] && echo "ERROR! Logging not initialised when calling get_log_files()." >&2 && exit 127 
+	[[ -z "${__logpath}" ]] && die "Logging not initialised when calling get_log_files()"
 
 	export OFS="|"
 	eval "$1=\"${__logfiles}\""
@@ -147,10 +159,10 @@ function get_log_files()
 # $2 should contain the recipient
 function send_logs_by_email()
 {
-	[[ -z "${__logpath}" ]] && echo "ERROR! Logging not initialised when calling send_logs_by_email()." >&2 && exit 127 
-	[[ -z "${__logredirfile}" ]] && echo "ERROR! send_logs_by_email() requires logs to be redirected first" >&2 && exit 127
-	[[ -z "${1}" ]] &&  echo "ERROR! send_logs_by_email() requires a subject as parameter 1." >&2 && exit 127
-	[[ -z "${2}" ]] &&  echo "ERROR! send_logs_by_email() requires a recipient as parameter 2." >&2 && exit 127
+	[[ -z "${__logpath}" ]] && die "Logging not initialised when calling send_logs_by_email()"
+	[[ -z "${__logredirfile}" ]] && die "send_logs_by_email() requires logs to be redirected first"
+	[[ -z "${1}" ]] && die "send_logs_by_email() requires a subject as parameter 1"
+	[[ -z "${2}" ]] && die "send_logs_by_email() requires a recipient as parameter 2"
 
 	mutt -s "${1}" \
          -a "${__logfiles}" \
@@ -171,7 +183,7 @@ function display_log_paths()
 # Removes all log files
 function clean_up_logs()
 {
-	[[ -z "${__logpath}" ]] && echo "ERROR! Logging not initialised when calling clean_up_logs()." >&2 && exit 127
+	[[ -z "${__logpath}" ]] && die "Logging not initialised when calling clean_up_logs()"
 
 	for lf in ${__logfiles}
 	do
